✅ Setup Database Connection (Not - Functional Yet) - Salvo
✅ Setup Database Controller for Models - Salvo

+++ InventoryManagementSystem2/Models/Customers.cs
@@ -9,7 +9,7 @@ namespace InventoryManagementSystem2.Models
         public int CustomerID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public BigInteger ContactNumber { get; set; }
+        public int ContactNumber { get; set; }
         [Required]
         public string Email { get; set; }
         public string Address { get; set; }
